Refactor context menu handling across character and grimoire modules
